# src/iron-pets/backend/src/api/openapi.yaml
openapi: 3.0.3
info:
  title: Iron Pets API
  version: 1.0.0
  description: |
    REST API for Iron Pets e-commerce platform - MVP version.
    Provides endpoints for authentication, product catalog, shopping cart, checkout, orders, and user management.
  contact:
    name: Iron Pets Engineering Team
    email: engineering@ironpets.com
  license:
    name: Proprietary
servers:
  - url: https://api.ironpets.com/v1
    description: Production server
  - url: https://staging-api.ironpets.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development server
...
security:
  - BearerAuth: []
paths:
...
  /checkout/validate:
    post:
      tags: [Checkout]
      summary: Validate checkout
      description: Validate cart items, stock, and address before payment
      operationId: validateCheckout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateCheckoutRequest'
      responses:
        '200':
          description: Checkout validation passed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateCheckoutResponse'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /checkout/create-payment:
    post:
      tags: [Checkout]
      summary: Create payment intent
      description: Create Stripe payment intent for order
      operationId: createPaymentIntent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '200':
          description: Payment intent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatePaymentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /checkout/confirm:
    post:
      tags: [Checkout]
      summary: Confirm order
      description: Confirm payment and create order
      operationId: confirmOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmOrderRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Payment failed or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /checkout/apply-promo:
    post:
      tags: [Checkout]
      summary: Apply promo code
      description: Validate and apply promotional code to cart
      operationId: applyPromoCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyPromoRequest'
      responses:
        '200':
          description: Promo code applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyPromoResponse'
        '400':
          description: Invalid or expired promo code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========================================
  # ORDER ENDPOINTS
  # ========================================
...
  /orders/{id}/cancel:
    post:
      tags: [Orders]
      summary: Cancel order
      description: Cancel order if status is pending or processing
      operationId: cancelOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailResponse'
        '400':
          description: Order cannot be cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{id}/reorder:
    post:
      tags: [Orders]
      summary: Reorder
      description: Add all order items to cart
      operationId: reorder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Items added to cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'

  # ========================================
  # USER ENDPOINTS
  # ========================================
...
  /pets/{id}:
    get:
      tags: [Pets]
      summary: Get pet details
      description: Retrieve pet profile
      operationId: getPet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Pet retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Pets]
      summary: Update pet
      description: Update pet profile
      operationId: updatePet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PetRequest'
      responses:
        '200':
          description: Pet updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetResponse'

    delete:
      tags: [Pets]
      summary: Delete pet
      description: Remove pet profile
      operationId: deletePet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Pet deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

# ========================================
# COMPONENTS
# ========================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authenticated requests
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request data
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
...
    Address:
      type: object
      required:
        - firstName
        - lastName
        - addressLine1
        - city
        - state
        - zipCode
        - phone
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        addressLine1:
          type: string
          maxLength: 255
        addressLine2:
          type: string
          maxLength: 255
          nullable: true
        city:
          type: string
          maxLength: 100
        state:
          type: string
          maxLength: 50
        zipCode:
          type: string
          maxLength: 20
        phone:
          type: string
          maxLength: 20
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AddressRequest:
      type: object
      required:
        - firstName
        - lastName
        - addressLine1
        - city
        - state
        - zipCode
        - phone
      properties:
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        addressLine1:
          type: string
          maxLength: 255
        addressLine2:
          type: string
          maxLength: 255
        city:
          type: string
          maxLength: 100
        state:
          type: string
          maxLength: 50
        zipCode:
          type: string
          maxLength: 20
        phone:
          type: string
          maxLength: 20
        isDefault:
          type: boolean
...
    Pet:
      type: object
      required:
        - id
        - name
        - species
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 100
        species:
          type: string
          enum: [dog, cat, small_pet]
        breed:
          type: string
          maxLength: 100
          nullable: true
        birthDate:
          type: string
          format: date
          nullable: true
        weight:
          type: number
          nullable: true
        weightUnit:
          type: string
          enum: [lbs, kg]
          default: lbs
        photoUrl:
          type: string
          format: uri
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PetRequest:
      type: object
      required:
        - name
        - species
      properties:
        name:
          type: string
          maxLength: 100
        species:
          type: string
          enum: [dog, cat, small_pet]
        breed:
          type: string
          maxLength: 100
        birthDate:
          type: string
          format: date
        weight:
          type: number
        weightUnit:
          type: string
          enum: [lbs, kg]
        photoUrl:
          type: string
          format: uri
    PetResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Pet'
...
    CartItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        productName:
          type: string
        productImage:
          type: string
          format: uri
        productSlug:
          type: string
        price:
          type: number
          format: decimal
        quantity:
          type: integer
          minimum: 1
        lineTotal:
          type: number
          format: decimal
        stockAvailable:
          type: integer
    Cart:
      type: object
      properties:
        id:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        subtotal:
          type: number
          format: decimal
        itemCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CartResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Cart'
...
    ValidateCheckoutRequest:
      type: object
      required:
        - shippingAddress
      properties:
        shippingAddress:
          $ref: '#/components/schemas/AddressRequest'
        billingAddress:
          $ref: '#/components/schemas/AddressRequest'
    ValidateCheckoutResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            valid:
              type: boolean
            stockIssues:
              type: array
              items:
                type: object
                properties:
                  productId:
                    type: string
                    format: uuid
                  requestedQuantity:
                    type: integer
                  availableQuantity:
                    type: integer
    CreatePaymentRequest:
      type: object
      required:
        - shippingAddress
        - shippingMethodId
      properties:
        shippingAddress:
          $ref: '#/components/schemas/AddressRequest'
        billingAddress:
          $ref: '#/components/schemas/AddressRequest'
        shippingMethodId:
          type: string
        promoCode:
          type: string
    CreatePaymentResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            clientSecret:
              type: string
            paymentIntentId:
              type: string
            amount:
              type: number
              format: decimal
    ConfirmOrderRequest:
      type: object
      required:
        - paymentIntentId
        - shippingAddress
        - shippingMethodId
      properties:
        paymentIntentId:
          type: string
        shippingAddress:
          $ref: '#/components/schemas/AddressRequest'
        billingAddress:
          $ref: '#/components/schemas/AddressRequest'
        shippingMethodId:
          type: string
        promoCode:
          type: string
        notes:
          type: string
    ApplyPromoRequest:
      type: object
      required:
        - code
      properties:
        code:
          type: string
          maxLength: 50
    ApplyPromoResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            code:
              type: string
            discountType:
              type: string
              enum: [percentage, fixed]
            discountValue:
              type: number
              format: decimal
            discountAmount:
              type: number
              format: decimal

    # ========================================
    # ORDER SCHEMAS
    # ========================================
    OrderItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        productName:
          type: string
        productSku:
          type: string
        price:
          type: number
          format: decimal
        quantity:
          type: integer
        total:
          type: number
          format: decimal
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orderNumber:
          type: string
        userId:
          type: string
          format: uuid
          nullable: true
        email:
          type: string
          format: email
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled, refunded]
        subtotal:
          type: number
          format: decimal
        shippingAmount:
          type: number
          format: decimal
        taxAmount:
          type: number
          format: decimal
        discountAmount:
          type: number
          format: decimal
        total:
          type: number
          format: decimal
        shippingAddress:
          $ref: '#/components/schemas/Address'
        shippingMethod:
          type: string
        trackingNumber:
          type: string
          nullable: true
        trackingUrl:
          type: string
          format: uri
          nullable: true
        promoCode:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        placedAt:
          type: string
          format: date-time
        shippedAt:
          type: string
          format: date-time
          nullable: true
        deliveredAt:
          type: string
          format: date-time
          nullable: true
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    OrderDetailResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          allOf:
            - $ref: '#/components/schemas/Order'
            - type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/OrderItem'
    OrderResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Order'